Code Challenge AgencyAnalytics
I developed a gallery application using React in accordance with the challenge guidelines. The libraries I utilized were React with TypeScript, React Redux, React Testing Library with Jest, and Styled Components. Additionally, I have written two tests for the GalleryItem and gallery slice. I have deployed the website online using Firebase. 

You are welcome to visit the website and explore it yourself.

You are welcome to visit the website and explore it yourself.